payment_service/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
celery[redis]
requests
django-cors-headers==4.3.1
django-redis==6.0.0